Execution update continuation
Given that code no longer depends on "spawn1" to load, removed all
references to spawn outside of actual code involving the spawner
(replaced with dependency on room). This includes changing both
room.init functions to depend on room instead of spawn as well.
NOTE: Still need to work on looping through a room's spawns instead of
depending on "spawn1" for spawning code.