Spawn Execution Update
Spawning code now loops for each spawn in room instead of depending on
"Spawn1". "saving" variable changed from spawn memory to room memory to
compensate.
Also changed room initialization method to execute in if-else if
statements instead of repeated if statements.
AFAIK next thing to work on is changing naming conventions to be
room-specific; and changing sources and source flags to be ordered by
proximity to spawns.