Completely Untested Code
This is an attempt to make the Harvester role act as all three roles in
one. I have done absolutely no testing of this code, so I have no clue
if it even works or not.
I also adjusted spawn.dict to heavily weaken the Upgraders and Builders,
and the Harvesters should mostly take care of those jobs.